corrections
- change of affiliation to "Scuola Normale Superiore, Pisa"
Van Kemseke publication date is wrong. It should be 2006 (instead of 2017)
- delete from notebook metadata (so it doesnt show in the bib, all references in text go to correct ref)
Costa, E. (2022). Socialist World from Balancing East and West to Militant Anti-Communism: Reading the Beginning of the Cold War in the Journal of the Socialist International (1947–1949). Journal of Cold War History, 24(1).
- change in abstract: "what role of European social democracy" to "what role European social democracy"
delete "text" in p30: "Topic modeling works under the assumption that the order of the words within the document text is irrelevant"
- change "Costa argued/s" to "I argue/d" in p12, 15, 101, 104 or "by me" in p6